shared-storage/shared-storage-permissions-policy-none.tentative.https.sub.html

<!doctype html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script src=/permissions-policy/resources/permissions-policy.js></script>
  <script src="/shared-storage/resources/util.js"></script>
  <script>
    'use strict';
    const same_origin_src = '/shared-storage/resources/shared-storage-permissions-policy-helper.html';
    const cross_origin_src = 'https://{{domains[www]}}:{{ports[https][0]}}' +
      same_origin_src;
    const header = 'permissions policy header shared-storage=()';

    promise_test(async t => {
      const allowed = await AreSharedStorageMethodsAllowedByPermissionsPolicy();
      assert_false(allowed);
    }, header + ' disallows sharedStorage in the current page.');

    async_test(t => {
      test_feature_availability('shared-storage', t, same_origin_src,
          expect_feature_unavailable_default);
    }, header + ' disallows sharedStorage in same-origin iframes.');

    async_test(t => {
      test_feature_availability('shared-storage', t, cross_origin_src,
          expect_feature_unavailable_default);
    }, header + ' disallows sharedStorage in cross-origin iframes.');

    const same_origin_writable_url =
      '/shared-storage/resources/shared-storage-write.py';
    const cross_origin_writable_url =
      'https://{{domains[www]}}:{{ports[https][0]}}' +
      same_origin_writable_url;

    promise_test(async t => {
      let response = await fetch(same_origin_writable_url,
                                 {sharedStorageWritable: true});
      let shared_storage_writable_header = await response.text();
      assert_equals(shared_storage_writable_header,
                    "NO_SHARED_STORAGE_WRITABLE_HEADER");
    }, header + 'disallows the \'Shared-Storage-Writable\' header to be sent '
              + 'for the same-origin shared storage fetch request.');

    promise_test(async t => {
      let response = await fetch(cross_origin_writable_url,
                                 {sharedStorageWritable: true});
      let shared_storage_writable_header = await response.text();
      assert_equals(shared_storage_writable_header,
                    "NO_SHARED_STORAGE_WRITABLE_HEADER");
    }, header + 'disallows the \'Shared-Storage-Writable\' header to be sent '
              + 'for the cross-origin shared storage fetch request.');
  </script>
</body>
